feat: move version information to settings.
logion-network/logion-internal#283
