+++ src/SettingsPane.css
@@ -1,0 +1,3 @@
+.SettingsPane .app {
+    margin-top: 30px;
+}
